Multi CPU cpi stack view

+++ tools/viz/viz.py
@@ -120,7 +120,7 @@ if __name__ == '__main__':
   elif num_intervals < defaultnum_intervals:
     # Automatically determine interval to end up with (around) num_intervals in total
     interval = defaultinterval * max(1, int(defaultnum_intervals / num_intervals))
-    num_intervals = defaultinterval * defaultnum_intervals / interval
+    num_intervals = int(defaultinterval * defaultnum_intervals / interval)
 
   if interval == None:
     interval = defaultinterval
